Refactoring, added Lombok, Service layer, get token by refresh token, added some todos
