+++ src/main/java/com/auth/testlogin/model/dto/TokenDto.java
@@ -1,5 +1,14 @@
 package com.auth.testlogin.model.dto;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class TokenDto {
 
     private String accessToken;
@@ -8,63 +17,4 @@ public class TokenDto {
     private String expires_in;
     private String scope;
     private UserInfoDto userInfo;
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getTokenType() {
-        return tokenType;
-    }
-
-    public void setTokenType(String tokenType) {
-        this.tokenType = tokenType;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-
-    public void setRefreshToken(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
-
-    public String getExpires_in() {
-        return expires_in;
-    }
-
-    public void setExpires_in(String expires_in) {
-        this.expires_in = expires_in;
-    }
-
-    public String getScope() {
-        return scope;
-    }
-
-    public void setScope(String scope) {
-        this.scope = scope;
-    }
-
-    public UserInfoDto getUserInfo() {
-        return userInfo;
-    }
-
-    public void setUserInfo(UserInfoDto userInfo) {
-        this.userInfo = userInfo;
-    }
-
-    @Override
-    public String toString() {
-        return "TokenDto{" +
-                "accessToken='" + accessToken + '\'' +
-                ", tokenType='" + tokenType + '\'' +
-                ", refreshToken='" + refreshToken + '\'' +
-                ", expires_in='" + expires_in + '\'' +
-                ", scope='" + scope + '\'' +
-                ", userInfo=" + userInfo +
-                '}';
-    }
 }
